Moved all the functions from the function folder to the helpers.....seemed redundent
