+++ helpers/generator/equipment/getFieldUpgradeList.tsx
