+++ helpers/generator/equipment/getLethalList.tsx
