+++ helpers/generator/equipment/getTacticalList.tsx
