+++ helpers/generator/getPerkList.tsx
