+++ helpers/generator/getStreakList.tsx
