+++ helpers/generator/getWildcardList.tsx
