+++ helpers/generator/weapons/getMeleeList.tsx
