+++ helpers/generator/weapons/getPrimaryList.tsx
