+++ helpers/generator/weapons/getSecondaryList.tsx
